reorganized components and debugged
