+++ two_windows_assignment/client/src/ComponentContainer/SingleInput.js
