+++ two_windows_assignment/client/src/InputContainer/CheckboxContainer.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import Checkbox from './Checkbox';
+import Checkbox from '../ComponentContainer/Checkbox';
 
 
 const items = [
